Fix nested projects (don't double-escape /)
The / is escaped by normalize_url() it seems, but that's not appropriate
for getting repositories in namespaces (e.g. sdk/git-lab becomes
sdk%2Fgit-lab, and then that % is escaped again in python-gitlab and
the repo isn't recognized). Unescape it, The 'safe' value for quote_plus()
is different from plain quote(), leading to / -> %2F conversions
(too early) which we don't want.